LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/local-attr-substitution-expected.txt


FAIL var() in attribute value substitutes locally assert_equals: expected "PASS" but got ""
FAIL var() in attribute value substitutes locally, argument assert_equals: expected "PASS" but got ""
FAIL var() in attribute value substitutes locally, typed assert_equals: expected "12px" but got ""
FAIL attr() fallback substitutes locally assert_equals: expected "PASS" but got ""
PASS attr() cycle through local
FAIL attr() cycle through unused fallback in local assert_equals: expected "PASS" but got "FAIL"
PASS attr() cycle through function
